backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic-settings==2.10.1
langchain==0.3.27
langchain-openai==0.3.28
langchain-qdrant==0.2.0
langchain-cohere==0.4.6
langchain-text-splitters==0.3.9
qdrant-client==1.15.1
openai==1.102.0
cohere==5.18.0
sentence-transformers==5.1.0
Pillow==11.3.0
textract==1.6.5
pypdf==6.0.0
python-docx==1.2.0
pandas==2.3.2
